MissionToMars
This is a submission repo for 2018 Swarmathon - Mission To Mars.

This implement involves multi-rovers to deliver targets from sub collection zones to the a main collection zone. A hector quadrotor is assisting the localization and navigation for the swarmie crew. The hector quadrotor is applied from a ROS package  hector_quadrotor.

The base code is provided by NASA Swarmathon.
Additional parts are done by the PCC-CSPU physical team's three main members.